requirements.txt
flask
psycopg2-binary
jsonschema
kafka-python
pip
setuptools